// src/types/error.ts
export enum ApiErrorCode {
  BadRequestException = "400000",
  ForeignKeyViolationException = "400701",
  UniqueViolationException = "400800",
  UnauthorizedException = "401000",
  DisabledUserException = "401600",
  InvalidUsernameOrPasswordException = "401700",
  TokenExpiredException = "401701",
  UserNotExistException = "401702",
  ForbiddenException = "403000",
  NotFoundException = "404000",
  RestrictionException = "409800",
  ParameterViolationException = "400700",
  DuplicateEntityException = "400900",
  UnhandledException = "500000",
  UploadImageError = "500100",
  UnmatchedBundleRuleException = "400702",
}